Improve commit message formatting with textwrap
The commit-msg.py script has been updated to use the textwrap library for
improved message formatting. The changes include importing the textwrap library,
and modifying how the final commit message is constructed by wrapping paragraphs
to 80 characters in width. This improves the readability of the commit messages
when displayed in various tools and platforms.
(gpt-4)